+++ src/app/app-routing.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { AuthGuard } from './auth/auth.guard';
 
 
 const routes: Routes = [
@@ -9,7 +10,7 @@ const routes: Routes = [
   },*/
   {
     path: 'editor',
-    //canActivate: [AuthGuard],
+    canActivate: [AuthGuard],
     loadChildren: () => import('./features/editor/editor.module').then(mod => mod.EditorModule)
   },
   {
