+++ src/app/shared/components/media-browser/media-browser.component.ts
@@ -2,7 +2,7 @@ import { Component, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { Observable } from 'rxjs';
 import { Media } from 'src/app/data/media';
-import { MediaService } from '../../../shared/services/media.service';
+import { MediaService } from '../../services/media.service';
 import { AppConfig } from '../../../core/config/app-config';
 
 @Component({
@@ -15,15 +15,15 @@ export class MediaBrowserComponent {
   title: string;
 
   get media$(): Observable<Media[]> {
-    if (this._data?.data?.filter) {
-      return this._mediaService.getMediaFiltered(this._data.data.filter);
+    if (this._data?.filter) {
+      return this._mediaService.getMediaFiltered(this._data.filter);
     }
     return this._mediaService.getMedia();
   }
 
-  constructor(@Inject(MAT_DIALOG_DATA) private _data: { data: { filter?: string; title: string } },
+  constructor(@Inject(MAT_DIALOG_DATA) private _data: { filter?: string; title: string },
               private _mediaService: MediaService) {
-    this.title = _data.data.title;
+    this.title = _data.title;
   }
 
   fixURL(file: string): string {
