+++ src/styles.scss
@@ -3,6 +3,7 @@
 @import '~bootstrap/scss/mixins';
 @import '~bootstrap/scss/utilities';
 @import '~bootstrap/scss/containers';
+@import '~bootstrap/scss/images';
 @import '~@angular/material/theming';
 @import 'theme';
 
@@ -260,7 +261,9 @@ app-nav {
     .item {
       display: flex;
       flex-basis: auto;
+      height: 4.3rem;
       order: 1;
+
 
       &:hover ul,
       ul:hover,
@@ -280,10 +283,10 @@ app-nav {
       }
 
       ul {
-        margin-left: -12px;
-        margin-right: -12px;
+        margin-left: -10px;
+        margin-top: 4.3rem;
         position: absolute;
-        top: -13rem;
+        transform: translate(0%, 50%);
       }
     }
   }
